# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestión Hoteles",
        "description": "Documentación de API para la gestión de hoteles, habitaciones y acomodaciones.",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/hoteles": {
            "get": {
                "tags": [
                    "Hoteles"
                ],
                "summary": "Listar todos los hoteles",
                "operationId": "88efeb5bd9555b55dbd8009d2a08d87a",
                "responses": {
                    "200": {
                        "description": "Lista de hoteles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Hotel"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error en el servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Hoteles"
                ],
                "summary": "Crear un nuevo hotel",
                "operationId": "05eb79906ed2370ac8a6b2dc4c37946a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "direccion",
                                    "ciudad",
                                    "nit"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Hotel Test"
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "example": "Calle Ficticia 123"
                                    },
                                    "ciudad": {
                                        "type": "string",
                                        "example": "Ciudad Ejemplo"
                                    },
                                    "nit": {
                                        "type": "string",
                                        "example": "123456789"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Hotel creado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Hotel"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/hoteles/{id}": {
            "get": {
                "tags": [
                    "Hoteles"
                ],
                "summary": "Mostrar un hotel específico",
                "operationId": "700f9f1a280d2ca8679b59b0455e6f8b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del hotel",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hotel encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Hotel"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Hotel no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Hoteles"
                ],
                "summary": "Actualizar un hotel",
                "operationId": "f9cfaba8f53e83675dea4cb9c71378f1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del hotel",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Hotel Actualizado"
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "example": "Nueva Calle 456"
                                    },
                                    "ciudad": {
                                        "type": "string",
                                        "example": "Nueva Ciudad"
                                    },
                                    "nit": {
                                        "type": "string",
                                        "example": "987654321"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Hotel actualizado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Hotel"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Hotel no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Hoteles"
                ],
                "summary": "Eliminar un hotel",
                "operationId": "04618d16a53b042046dbc03601affeb7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del hotel",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hotel eliminado con éxito"
                    },
                    "404": {
                        "description": "Hotel no encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Hotel": {
                "required": [
                    "id",
                    "nombre",
                    "ubicacion",
                    "descripcion"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Hotel Paraíso"
                    },
                    "ubicacion": {
                        "type": "string",
                        "example": "Calle Ficticia 123"
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Hotel de 5 estrellas con vista al mar"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}